move some zones data structures around
